vtkCutter: Add new option to produce contour polygons
The polygon slicing is turned on by setting an additional parameter
called OutputTriangles to be false. Since it is on by default,
behavior of existing code will not be affected.

The main addition is a utility class vtkPolygonBuilder that
incrementally maintains the boundary of a triangulation as it grows
one by one. It is used during all slicing to compute the polygon when
slicing a single input cell.

Change-Id: I18dd45efc790f13610f4a839839a7eb256659111