Fix floating-point vtkVariant comparison and test.
In theory, all of these variants should compare as equal:
   vtkVariant flt(0.583f);
   vtkVariant dbl(0.583);
   vtkVariant str("0.583");
Currently, they do not because the 32-bit float is promoted
to a 64-bit double instead of the other way around.
This commit demotes doubles to floats when one operand is a float
and tests the result.

It also enables the `TestVariant.cxx` test, which was in the
repository but not being built or run.

Change-Id: I9f2c2cc04ff1ee6961b92f44045c54797e5fe011