COMP: Fix compile problems due to empty VTK_TK_LIBRARIES.
The VTK_TK_LIBRARIES variable is set before VTK_USE_TK is set,
but because of a conditional this caused VTK_TK_LIBRARIES to
never be set on the first cmake run.  The VTK_TK_LIBRARIES
should be set whenever the Tk libraries are found.

Also fix forced inclusion of renderingtcltkwidgets package in
pkgIndex.tcl, it should only be include if it was built.

Change-Id: I7082a3440635e7927423ab42f497bbef6bf56bbf