Allow Grid visibility to be independant of the axis visibility
This is specially usefull in closest and furtherst mode when you
only want to see the grid on the back regardless the axis location.

For instance the axis could be on all the edges, but now the grid
can be shown either everywhere or only on the front or the back.

Change-Id: Idb382c08f459632a8591964dfd4abde62ea991b1