refactor vtkChartXYZ
It now follows the API of the 2D charts more closely.  vtkChartXYZ is
responsible for the axes, while the subclasses of vtkPlot3D handle
displaying the actual data.  As part of this effort, I've included
optional interactivity into vtkChartXYZ.  This eliminates the need for a
separate vtkInteractiveChartXYZ class.

This change also introduces new functionality.  vtkPlotSurface allows us
to visualize a table as a 3D surface plot.

I've updated vtkScatterPlotMatrix so that it correctly uses the new API
of vtkChartXYZ for animation.

All affected tests were updated as well.
Change-Id: Ic8406c99758a98851949c6153129f7704784e31a