Make `vtkVariant::ToNumeric()` private.
Although `ToNumeric()` was exposed as a public, templated method,
using it outside of the vtkCommonCore library causes link
errors on Windows. This commit makes the method private and
switches other classes to use either the templated `vtkVariantExtract`
function or the `ToXXX()` methods already present in `vtkVariant`,
depending on how `ToNumeric()` was being used.

Change-Id: I6c7f3f1ebb13d0d5ad4137dfeb4753c55c28a88d