Fix a compilation issue when crosscompiling.
The issue appears when crosscompiling on Windows using NMake and
targeting the Android platform (Linux & gcc based).
The vtkmkg3states tool is not built (because of the flag
CMAKE_CROSSCOMPILING), however, the tool is still called to
generate the tif_fax3sm.c file, which leads to a unknown file error.

I do not know the internals but tif_fax3sm.c already exists and
to fix the issue I propose to call the vtkmkg3states tool only
if it has been built.

Change-Id: I924c39bf692fa005e8852ddb6a4b8a7f2c3ed6a8