CTestCustom: Fix signed-overflow warning regex syntax
The parent commit added a regex with invalid syntax.  Not only does it
not match the desired warning but it causes CTest to print an error when
compiling the regex.  The error message may appear in custom command
output when CTEST_USE_LAUNCHERS is enabled.  For files generated by
redirecting custom command output to a file the error message then
appears in the file.

Change-Id: I77922d165e5ad7ec38ca33a5915b76b455470ed5