BUG: Depth peeling not working on MAC
The following tests fail on the MAC:
TestGaussianBlurPass
TestTranslucentLUTDepthPeeling
TestTranslucentLUTDepthPeelingPass

They all use vtkDepthPeelingPass.
On MAC's, vtkOpenGLProperty reports the error:
Samplers of different types use the same texture image unit.
 - or -
A sampler's texture unit is out of range (greater than max allowed or negative).

Logic in vtkDepthPeeling::RenderPeel binds three textures. If either
OpaqueShadowTexUnit or ShadowTexUnit is 0, the

vtkgl::ActiveTexture(vtkgl::TEXTURE0+this->OpaqueShadowTexUnit);
or
vtkgl::ActiveTexture(vtkgl::TEXTURE0+this->ShadowTexUnit);

will be the same as

vtkgl::ActiveTexture(vtkgl::TEXTURE0);

This can generate the condition for "Samplers of different types use
the same texture image unit.".

The solution is to avoid using texture unit 0 for either of the
classes textures.

Change-Id: Ib6217ca70de63d1c7e9e281b346294e05c6db0af