REMOVE-TEST: removing vtkImagingCorePython-WindowLevelInterface
The corresponding tcl test is not enabled. It is reasonable to
assume that this is legacy test. We will keep the code in the
repo in case somebody else finds use for it and can repurpose it.

NOTE: If one want to fix this test then they would also need to
enable the corresponding TCL test and make that work as well.

Change-Id: Ic1af6ac702535714b00040de91854d19b28151ee