various improvements for vtkTreeHeatmapItem.
1: Improved rendering performance by only drawing data that
is visible on the screen.  This is important for the visualization
of large data.

2: Improved visualization of collapsed subtrees.  We now use color
to indicate how many leaf nodes are hidden within a collapsed subtree.

3.  More changes to collapsible subtrees.  You can no longer collapse
a single leaf node.  This required a slight change to the test to
ensure that the relevant code remains covered.  I also made sure that
the visualization of this item does not change when a subtree is
collapsed & re-expanded.  This requires me to update the baseline
for TestTreeHeatmapItem

Change-Id: Ib750c6056939b61b11e693f910dcbcca91217868