Fixing POutlineFilter and POutlineCornerFilter for CompositeData types.
Updating vtkPOutlineFilter and vtkPOutlineCornerFilters to work correctly with
all types of composite datasets in parallel. Based on the type of the dataset
we are dealing with, the filters now reduce the bounds, as needed and produce
outlines. Based on the data type, the outlines may be produced on the root node
(for multiblock datasets, overlapping amr or vtkDataSet) or on all nodes (for
  non-overlapping amr).

Change-Id: I761f8570db39307fb767607fdebae2bafab71af4