Change python arg float-to-int conversion to python 2.7 style.
In Python 2.3 through 2.6, if a method that expected an integer
parameter was given a float, then a DeprecationWarning was printed.
In Python 2.7, a ValueError exception occurs instead.  This patch
causes the vtk python wrappers to follow the python 2.7 behavior:
If a VTK method expects an int but is given a float, an exception
will occur.

The rationale for this change was that it is very difficult to find
out what python code is causing the "deprecation warning", because
it does not print the method or line number.  If an exception is
generated, the the user sees a stack trace and can therefore easily
fix the python code.

Change-Id: I11f0efb506884fc4eedae84bb251b10102daae48