BUG #11923 fix non power of two framerates
ffmpeg is picky about the bit rate tolerance
At least with ffmpeg 0.6.5, it won't make a codec if brt is < bit rate
(including the default value of 4000000) unless you have a power of
two framerate in some cases for some strange reason.

Setting brt = br appears to be a reliable choice. Perhaps because
this gives ffmpeg free reign to very as much from the target as it wants.

Change-Id: I6100acd21bf979c4d7b7448f4e32b26d2de7cdf7