Make sure the heatmap is rendered when it should be.
This commit fixes a bug that caused the heatmap to not be rendered
when it should be.  The problem was that this class would count
collapsed trees contained by other collapsed trees when deciding whether
or not any leaf nodes remain expanded.

As of this commit, we only count a collapsed tree if its root is still
part of the pruned tree.  We still have a heatmap to render if the number
of such collapsed trees is less than the number of leaf nodes.

Change-Id: I1a21940e651f5fd59f3e9a4c785c8c8eedc529f0