Fixing Clear() method for the context scene
Clear() was removing elements from the scene without first
setting the scene and parent to NULL, as is done in
RemoveItem(). This sometimes caused a crash in e.g.
vtkContextScene::MouseMoveEvent() where it keeps a reference
to the last item picked and attempts to invoke a method on
it as long as the scene is still valid. This change causes
this check to not pass (scene is NULL) and avoid the crash.

Change-Id: Ic9a64f38b5c9093ea6e56241ba16e4bba7cb99ca