Exhaustive and sampled distinct-value detection.
This change renames "unique" value detection to "prominent"
value detection, since we now allow both exhaustive detection
of every unique value in an array as well as a statistical
approximation to this which may miss rare values but for which
we have a statistical parameterization of the confidence that
non-rare (i.e., prominent) values will be detected.

Thanks to Sesh Comandur, Janine Bennett, and Ali Pinar (Sandia
National Labs) for the parameterization of sample size by
uncertainty and rarity.

Change-Id: Ia35dd897be18bada177184515e18bcb9298360ff