Add a TextRenderer interface and implementation.
An abstract vtkTextRenderer interface is added to RenderingCore.
This will allow all rendering classes to have access to text rendering
functionality without an explicit dependence on RenderingFreeType. Once
existing code is migrated to use this simplified interface, applications
using VTK will be able to add their own text rendering backends easily.

The interface is designed to, at minimum, switch between FreeType and
MathText rendering, enabling any users of the interface to have access
to pretty-printed equation by passing in MathText markup, which is a subset
of LaTeX. The implementation will test for the presence of "$...$" in the
input string and, if present, MathText will be used instead of FreeType.
If MathText is unavailable (it requires python and matplotlib) or fails
to render, a warning is produced and the FreeType backend is used as a
fallback.

The default implementation of this class is vtkMathTextFreeTypeTextRenderer
in the RenderingFreeType module.

Change-Id: I215ad856e3f197c3ea22eb8f9dd68e5ec2f9e620