Allow an image reader to read data from buffer instead of file.
This feature may be very useful to avoid the serialization of a temporary
image before reading it with a VTK reader - for instance an image fetched
from an internet socket or any procedural fashion...

The 2 buffer related variables (MemoryBuffer and MemoryBufferLength) are
implemented in vtkImageReader2.
For now, only the JPEG reader supports this new feature.

Change-Id: I9d00e19c92143cd4190353ccabf574012696d6db