Adding mechanism to locate Hierarchy.txt file for a module.
This patch adds a variable to a module-config file that identifies the location
of the Hierarchy.txt file. Thus projects no longer have to guess the location
for a Hierarchy.txt file, it is imported when the module's configuration is
imported.

Note that we are still not installing Hierarchy.txt files for development
installs, hence the locations for the Hierarchy.txt file in the installed
module-configs is currently bogus..

Change-Id: Ibaa0324765753f60e5a15e319f2367d75ce7c5c8