BUG: No data for TestFlashReader
TestFlashReader requires input called: smooth.flash.

According to Yuanxin Liu:
This data set (smooth.flash) is big (1.7G), which is probably why no
one has tried to commit it to VTKData. Unfortunately, as far as I
know, it is the only flash data set that is passed around among the
developers here.  If anyone has a small .flash data set that is good
for testing.  Please post it !

Until a smaller dataset is fouond, remove this test.

Change-Id: I05de065256215f31a7e982ccd7737488ece84ca0