Adding CMAke options to select from available DAX_BACKENDS.
Currently Dax can use four backends namely
 - CUDA
 - TBB
 - OpenMP
 - Serial

Dax may be build with only a few option. So when building with a Dax
build only options with which dax was built will be available. However
the dax acclerator in VTK cannot use all available options and should
choose only one of the availabe options.

This commit enables the user to pick from available options in the DAX
build. The default order of preference is CUDA > TBB > OpenMP > Serial.
A higher preference one is selected by default only if available.

Change-Id: I360b3d6cdeea0be93108b98d8dee01ca78425303