Fix thread-safety issue in vtkImageData::GetDimensions()
This fixes a thread-safety issue in the vtkImageData::GetDimensions()
method. The thread-safe version (which takes a pointer to the output
array) was calling the non-thread-safe version causing both to not be
thread safe. The calculation of the dimensions was moved to the thread-
safe version so it can now be reliably called from multiple threds.

Change-Id: I156d6e196c5b7a0ab4654941a646dfad4e4cfefe