ENH: Reduce grid size to 10x10x10
Changed whole grid extent to be a 10x10x10 grid in order
to reduce the memory footprint of this test, which was
causing failures on the AIX dashboard. The test now
partitions the grid to 4 blocks and uses 1 layer of ghost cells
to check the integrity of the ghost-layer/ghost-array generation.

Change-Id: Ibb4b8418bd3c8b683b342b4c08f0e5753e37a331