Initialize ivars to 0 to prevent access of uninitialized data
With -fsanitize=signed-integer-overflow, the test
'vtkIOXMLPython-TestEmptyXML' gives:

IO/XML/vtkXMLPStructuredDataReader.cxx:62:35: runtime error: signed integer overflow: 591396864 * -536868871 cannot be represented in type 'int'

The reason being that 'PointDimensions' was never set and so contained
random values that happened to be pretty big. There probably should
be some better error checking somewhere, but it seems a good
defence to zero these things anyway, in case a client of the class
(incorrectly) invokes a method that accesses the uninitialized data,
at least we'll get consistent results.

Change-Id: I77f5ac68d4edad632931667ddf5e87f930f48b3f