Fixed the 'double' variant of SetColor to use 1.0 not 255
I believe there was a copy-paste here.  The alpha was being
set to 255 (notably not written 255.0) instead of 1.0 for
the 'double' case, where rgba values should (I believe)
range from 0.0 to 1.0.  Fixes undefined behaviour reported
by clang:

runtime error: value 65025 is outside the range of representable values of type 'unsigned char'

that happens when this 255 value was itself later multiplied
by 255 in an attempt to change it to uint8 range.

Change-Id: I85c718bcbc275a328bb6264222d4dec2c3110781