Fix python.h induced Visual Studio compiler warning
Python.h defines hypot to be _hypot, which can cause a level 4
warning coming from math.h with the Visual Studio compiler.
Use pragma warning to make sure that only level 1 warnings
may be emitted during the include of Python.h.

Also, remove the custom expression that was trying to suppress
this warning since it had a missing space in it that prevented
the match anyhow. It's now unnecessary since the warning is
not emitted.

Change-Id: Iaf77a0ae76ebab5056fb02d3895e2084168ad731