Add reader/writer for NIFTIv1 and NIFTIv2 files.
This patch adds support for reading and writing NIFTI files, as well
as full access to the NIFTI header.  Versions 1 and 2 of the NIFTI
header are supported, and the reader can also be used to read Analyze 7.5
files, though the only Analyze header elements that are supported are the
ones that overlap with NIFTI.  Automatic compression and expansion for
gzipped files is also provided whenever the filename ends with a .gz
extension.

Change-Id: Iead3afb1a65d8c01aa52f3ca55c6da2f08703375