fixed out-of-range stack buffer read
Found by clang's Address Sanitizer.

The 3rd element of a stack array of size 2
was being read, which is out of range.

After looking at the code and usage,
I believe the size of the array is correct,
and that the out-of-bounds read is what's
wrong.

Instead of always reading garbage,
I've substituted the value 1.0.

There may be a better fix…

Change-Id: Ib270d87dc98bfd66035f48e0d5718f7862f86cef