Expose the vtkExodusIIReaderPrivate::SqueezePoints variable.
We need to pass the full node data into the reader's output
to do a comparison with the in-situ reader. The SqueezePoints
variable controls whether or not to prune points that don't
belong in a given element block.

This patch simply adds methods to vtkExodusIIReader that
set/query the private class's SqueezePoints variable.

Change-Id: If6ceaa11939abfd731c4c172bf305b74f5c3878d