Optimize HTG for large & sparse level 0 grids.
- Replace the vtkCollection which has a find access in O(N) by a std::map.

- Add vtkHyperTreeIterator to avoid browsing every potential tree of
level 0 grid. Only available trees are now browsed.

- Add a TimeStamp to avoid unnecessary calls to the costly
UpdateHyperTreesLeafIdOffsets.

Those 3 changes really improve the computation time for building the HTG
but also for computing the dual and all the filters
(AxisCut, Geometry & ToUnstructuredGrid).

Also add points for each cell in the tree, not only for leaf, even if
only leaf points are used to construct the dual and thus referenced by
some cells.

Change-Id: Ie954fdc8860a55d0de9c23097b0662a2d2123550