Attempts to fix random failure of TestPolygonSelection on amber8.
The original test has many MouseMove events, and each event will trigger redrawing of polygon on screen,
which takes a fairly long time to run on amber8 due to software-rendering with Mesa. This commit modified
the playback events so that only a few events are included and the test result image stays the same. The pixel drawing part of the test is also turned off for the test.

Change-Id: I848a550ad4a3e0fd662fb97df96e381587585318