Fixed bug in vtkCachingInterpolatedVelocityField.
vtkCachingInterpolatedVelocityField was passing -1 to GetCell() the
first time around. Since GetCell() expects a valid cell id, this was
causing all kinds of issues. Also, it was not calling FindCell()
the right way - it was always passing -1 for the cell id causing it
to use the locator every time. Fixed both.

Change-Id: I8025666e046f4b39d7b43d5c3b71f69f763b20ef