Use the font scaling provided by vtkTextActor.
This will increase the size of fonts used on displays
with high pixel densities.

Also, this changes the signature of vtkTextActor::GetBoundingBox()
to take a vtkViewport. Since this is a new method for VTK6, it is
better to fix it now; having it pass a NULL viewport pointer
internally was causing crashes.

Finally, vtkTextActor::SetConstrainedFontSize() should take font
scaling into account.

Change-Id: I3cc99f99798da96f5c8e9249538ba269216e302a