change leaf node display behavior
Prior to this commit, leaf nodes were extended so that they all
lined up vertically.  The motivation behind this behavior was
to make it easy to see which leaf nodes correponded to each
row in the heatmap.

The problem with this approach is that it obscured information
about how long these leaf nodes actually are.  Starting with this
commit, the actual length of the leaf nodes is now drawn in black
by default (the same as the rest of the tree).  The leaf nodes are
still extended as before, but this extra length is now drawn in grey.

This extension of the leaf nodes is also now a toggleable behavior.
The options are always extend leaf nodes, never extend leaf nodes,
or only extend leaf nodes if a heatmap is also loaded.  This latter
option is the default.

This commit also introduces two new public member functions: GetSize
and GetCenter.  These functions make it easier to center this item
within a scene in an end-user application.

Change-Id: I11576b8822de2c915a6edb36c80c4721661aa317