separate color maps for categorical & continuous data
This commit separates the default colors that we use to represent
categorical & continuous data within our heatmap.  Continuous data
is now drawn using a blackbody color map.  Categorical data is drawn
using the diverging accent color set.  In order to take advantage of
vtkColorSeries in this effort, a dependency on vtkCommonColor was
added to this module.

As a result of this effort, we now use one vtkLookupTable to
assign colors to all continuous data, and another for the categorical
data.  This is a simplification over our previous approach, where we
used a separate vtkLookupTable for each column.

Change-Id: If008b871c750b1020b8df04c89e34db36785f853