change MAX_DISCRETE_VALUES into an ivar
Change MAX_DISCRETE_VALUES from a hardcoded value of 32 to a user-
settable ivar.  This allows users of vtkAbstractArray to generate
a categorical color map for data sets containing more than 32
discrete values.

For backwards compatibility, we keep the MAX_DISCRETE_VALUES enum
and use it to initialize MaxDiscreteValues.  This enum is marked
as deprecated.

Change-Id: I83bf7c598288dc28a4ad231778c66e1571e3a461