Fixed regression introduced by a6fa3e375c.
a6fa3e375c changed the logic in vtkWin32OpenGLRenderWindow::CleanUpRenderers().
That method was used when swithcing from offscreen to onscreen. That resulted in
the renderers not having correct references to render window after switching
from offscreen to onscreen context causing dashboard failures on ParaView
dashboards. This fixes that issue.

Since the a6fa3e375c changed the logic in CreateOffScreenDC() by not using
CleanUpRenderers() to reset renderers, it makes sence to change the logic in
ResumeScreenRendering() to not use CleanUpRenderers() for resetting renderers as
well.

Change-Id: I32e58854fa8720c99212cee5d7a2652f11dc9b21