Split Python module initialization function out
The module initialization function has always been put into the Python
module itself which is marked as a MODULE in add_library in the VTK
wrapping macros. CMake disallows linking directly to a MODULE, so if
something external to the bindings want to use the module initialization
function, they don't have access to it.

To fix this, the module method is placed in an OBJECT library which can
be linked against to get access to the init function. In addition, the
code which loads each class into the module is moved into the PythonD
library. The Python module itself doesn't use the object library since some
generators do not support targets with only objects from other targets.

This is mainly intended for use in ParaView so that the Python bindings
for classes may be used to implement the ClientServer wrappings with the
Python wrappings instead of generating a second set of bindings.

Change-Id: I8dd2eb672a3b038dedafa79cae271b88fad06082