Properly hide the subclass from the wrapping code
Other code as tried to follow the pattern used in VTK for masquerading
one class as the superclass instead of another for wrapping purposes.
Unfortunately, the way VTK does it isn't a good example since
vtkDataArray is actually a special case in the parser VTK uses for
generating the bindings. Using the __WRAP__ symbol the wrapping code
defines is a better way of handing this.

The wrapping grammar is also updated to not tokenize vtkDataArray.

Change-Id: I9dd9b2d8d3b98c70e3c1542608690ead46355203