Fix dependency issues with vtklibproj4.
vtklibproj4 was copying the vtk_libproj4_mangle.h file to output on every build.
That lead to cyclic dependencies and the build was always dirty (especially when
using Ninja). This fixes the issue by removing the unnecessary copy and ensuring
that the include dirs are setup correctly to include the vtk_libproj4_mangle.h
header.

Change-Id: I86d7972c118a53265e8240aea779413bf6ce0f64