Fixed finalization bug in vtkTimeStamp.
Finalization of vtkObjectFactory was calling vtkTimeStamp::Modified
after the static atomic int was destroyed. Used Schwarz Idiom
to fix the initialization dependencies.

I also had to change ProcessShader.cxx to not include vtkObject.h
as it does not link against vtkCommonCore. All it needed was to
include system headers to get iostream.

Change-Id: I9c5959e6e17c7c307f0d2014444591b197e82e19