Add Graph Layout example in VTK-Web
The example generate a random graph and display it
within the Web browser window in full screen.
The user can also switch from several layout by
changing the selected value in the combo-box on
the top right corner.

To run it, execute the following command line from the build tree:

$ ./bin/vtkpython Wrapping/Python/vtk/web/vtk_web_graph.py --content www --vertices 10 --edges 10

Change-Id: I9072afa3896c48de7cdaa89857b77773234db06e