Fix MSVC 2008 warning about valid STL usage.
"warning C4996: 'std::copy': Function call with parameters that may be
unsafe - this call relies on the caller to check that the passed values
are correct. To disable this warning, use -D_SCL_SECURE_NO_WARNINGS.
See documentation on how to use Visual C++ 'Checked Iterators'".

Because the std::copy destination is a raw pointer.

Some solutions to get std::copy to work with pointers:
http://choorucode.com/2010/08/30/visual-c-c4996-warning-on-copy-with-array-parameters/

So yeah...Let's go with memcpy.

Change-Id: If0ff6277625cf1f63535bec72598587d4b934afa