Fix bad point detection for large double values
The bad point detection was incorrectly detecting large double values
as bad, use templated code that works with the original type and a set
to ensure a sorted, unique list of bad points is generated. Modified
the test to use a second corner and one series that is very large to
test the functionality.

Change-Id: I08322a0b3fd728e496d2d5f1f1901054afe4eb7d