Added CellMask support to the MPAS reader.
MPAS reader can now read cell masks to blank out land cells.
Use Threshold to remove cells that have a CellMask of 0.
Note that the variable is actually called vertexMask in the
file. This is because we are loading the dual (triangle) grid
where vertex (primary grid) == cell (dual grid).

Change-Id: Ia6171b932de338fac8f0622f2c8dc84ab4f592c6