Wrap methods with POD pointer arguments in Python
Support wrapping parameters in Python which are pointers to POD types.
These parameters are passed using memory managed by a vector of the base
POD type (except bool since its memory layout is not compatible with
bool* semantics) so that no memory is leaked. The size is obtained from
the passed-in argument, not any size argument which the C++ code may
use, so Python code must be careful to match where necessary.

Change-Id: I8c85baf76374ecaa5a65348c998db8f56941dc86