improve appearance of VTK web example
In this commit we change some of the settings on our
delimited text table reader.  This results in improved
appearance for our phylogenetic visualization.

We tell the table reader to detect numeric columns so
our visualization will know to use the continuous data
color scheme for these columns.

We also tell the table reader that our input .csv file
has column headers.  This results in the correct column
labels being displayed in the resulting visualization.

Change-Id: I9d5e8a13a6f705c102df6fa2ba0f9cd5734cbab8