Make cell blanking work as expected in structured grids
This change makes it so that the default rendering path will
properly hide blanked cells. When surface filter is given
a structured grid with blanking, it now takes the slow
generic path because blanking will expose interior faces.
In there, it now skips blanked cells and treats faces of
cells that are shared with blanked cells as exterior faces.

Change-Id: I23d6397f8f9c7bbda3c366d160eaa03c4668be48