Fix failing printOther test.
Older windows compilers created code that will segfault when a NULL
char pointer was passed to an ostream, as they were interpreted as
strings. When the length of the string was calculated, the NULL check
was omitted, causing a segfault when 0x0 was dereferenced to check for
'\0'.

This patch casts the pointers to void* so that the pointer value will
be printed.

Change-Id: Ice4339d14205154a27653bd14d9f955054a01f55