Improved performance by removing unnecessary indirection.
The subclasses of vtkDataArrayTemplate needed to declare and
forward to superclass all methods defined in vtkDataArrayTemplate.
This was so that the wrapper code could wrap these methods
because it currently skips the templated superclass, which
can't be handled by Tcl and Java generators. This redirection
added an unnecessary performance hit. I changed the code such
that these methods are declared for wrappers' eyes only. This
way, there is not performance hit but the wrappers see them
still. I also moved this code to a preprocessor macro to make
it easy to maintain.

Change-Id: I88e12d00a49a00b431110c69aabbec0fa7d689c7