Hide SetArray method on vtk*Array classes from wrapping
These methods use the pointer given directly. Since wrappers would have
to know this and keep the converted array available for the class. It
would also be out-of-date with the actual language-end array (which
might even be ephemeral, so what keeps track of ownership in that
case?). It's best to just hide these methods since improper use is
almost guaranteed to cause a segfault, not an error message.

Change-Id: I276b4a5332beddf027e865b2dc2e158311ed7587