BUG: Fixed a bug introduced by removing redirection from arrays.
In a previous commit, I made sure that vtkDataArrayTemplate<>
subclasses don't have to define methods that forward to
superclasss just for the sake of wrapping code. One issue this
introduced was that in vtkIdTypeArray, GetDataType() was
returning VTK_ID_TYPE. But the superclass methods returns
the native type of vtkIdType which may be int or long long.
This is because the templated code cannot be specialized for
a typedef that resolves to a native type. Fixed the issue by
re-introducing GetDataType() to vtkIdTypeArray. Note that
by definition vtkDataArrayTemplate<vtkIdType> != vtkIdTypeArray.

Change-Id: I684e0901be8e37522d00d6a50d2cfb598919e446