Added code to handle NULL whole extent in RequestUpdateExtent().
When structured data filters are used in a composite data pipeline,
it is possible that WHOLE_EXTENT() will not be available the
first time RequestUpdateExtent() is called. Filters that work with
the composite pipeline need to handle this. Fixed for
vtkSynchronizedTemplates3D. Will re-evaluate a broader fix in the
future.

Change-Id: I24051df9a132efa8409aa99cebf37e34b707f14d