Add parsing of variadic templates for C++11 support.
This was done by adding opt_ellipsis where it was needed for C++11
(specifically after base class IDs and before declarator IDs), and
then adding opt_ellipsis in a whole bunch of other places in order
to eliminate shift/reduce errors in the parser generation.  Base
class specifiers that are parameter packs are ignored for now.

Change-Id: I86d13da5ed6585a529e07434946b33654a54e9a9