Eliminating CMP0020 warnings in the examples.
If VTK is built with Qt5 and CMake >= 2.8.11 CMP0020 warnings are generated
 in these examples. Setting CMP0020 to new eliminates these warnings.

This warning happens because the examples link to ${VTK_LIBRARIES}
which, inside the VTK build tree, is all libraries built including
those that use Qt.  The policy is set as a workaround in these cases.

A future submission will modify these and other examples to just
use the needed libraries.

Change-Id: Iea460773065ed55880b198eeaeec3da3b711c258
Signed-off-by: Andrew Maclean <andrew.amaclean@gmail.com>