Adding support for multiple inputs/outputs
Allow the vtkRCalculatorFilter to have multiple inputs/outputs
of type tree or table.

Add four methods to API to achieve this:

  void PutTrees(vtkStringArray* NamesOfRvars);
  void GetTrees(vtkStringArray* NamesOfRvars);
  void PutTables(vtkStringArray* NamesOfRvars);
  void GetTables(vtkStringArray* NamesOfRvars);

These methods allow string arrays to be provided to set multiple
inputs or outputs. The vtkTrees or vtkTables are then assigned
to these names in the order in which they appear in the input or
output dataset.

Change-Id: I1a7049619502435666ec0dc1e6fe2a8ebfc479c9