Added Doxygen support for configured files
Include the binary directory, in addition to the source directory,
for Doxygen to consider. The source directories are still entirely
independent of the build configuration, architecture, etc. The
addition of binary directories will consider generated files, which
will only be generated if the module is enabled and may be affected
by various build, compiler, and configuration parameters.

Change-Id: If2a44d9628ff57fb1b74d4a3d4553bd11ba5ba08