Initialize unused variables when setting event
The vtkContextMouseEvent object isn't really designed to be used
outside of the mouse events/interactors. When using them all variables
should be set. Ensure positions that happen to not be used in this
particular test are initialized (this is all normally taken care of
when real mouse events are received).

Change-Id: I12ba843845e42dddb7e2ddb5ef05492b79f10786