Add a vtkweb module providing application code for running web tests.
The new module allows addition of vtk (and also paraview) tests and
gives a simple interface for selecting the tests to be run and then
running them.  Output includes an error log which can be retrieved
by Python/selenium tests or viewed in the web application.  Also
included is a simple list view of which tests have passed and which
have failed.

When testing is enabled in cmake, the new module is included in the
core set of vtkweb javascript modules automatically.

Change-Id: I95469a6a8a54d6d786220a8a4ff411acaf3a7784