Enhance WhatModulesVTK.py
The python script accepts either a file as the second
parameter or a directory.
If a directory is passed to the program it is
recursively searched for all files that contain includes.

The script has been enhanced to parse the module.cmake files
in order to calculate, in addition the set of all modules
referenced in the files, the following extra sets:

1) A minimal set of modules that need to be referenced.
2) A set of modules where all the dependencies are chased down.

Change-Id: Ie0251a76c85458ddb8a1ed034d3d89239abddcae
Signed-off-by: Andrew Maclean <andrew.amaclean@gmail.com>