only create an edge weight array when necessary
Change vtkNewickTreeReader so that it no longer creates an edge
weight array if the input file does not contain edge weight information.
This makes it easier for vtkNewickTreeWriter to detect whether or not
edge weights should be written to the output file.

This commit also contains corresponding changes to vtkDendrogramItem.
This class no longer assumes that the input tree has an edge weight
array.  As of this commit, vtkDendrogramItem also allows users to
specify the names of the vertex weight & vertex name arrays within
their input tree.  Previously, these array names were hardcoded to
"node weight" and "node name", respectively.

Change-Id: Icdfe6c641c2257686f647865b42f4f4160c6d179