Regenerate lex.yy.c and vtkParse.tab.c for C++11 support.
The changes made to lex.yy.c and vtkParse.tab.c in this patch are
inclusive of all changes to vtkParse.l and vtkParse.y since gerrit
change I80b692a129677cb3f1ccfb7177c4c7099440ec72.

The changes were tested by generating the wrapper source for Python,
Tcl, and Java and performing a diff to ensure that they were identical
to a copy of the wrapper source that had been generated before applying
the changes.

Change-Id: I139d707d80d277ca727fc367aad0a607c6053bc5