fix failing RCalculator tests
This commit resolves some issues that lingered after 5bd3962.
There were still a few spots in vtkRAdapter where we compensated
for an extra root node.  This is now fixed.  The tree tests in
this module also had to be updated now that VTK trees & R trees
contain the same number of vertices and edges.

Change-Id: I624c6e23b08146f91d3141aded8ccf380d9f6538