display scalar bar for colored tree
vtkDendrogramItem now displays a vtkColorLegend after SetColorArray()
has been called.  This allows the consumers of this visualization to
see the scalar values associated with the colored edges of the tree.

This commit also contains a bug fix.  When SetColorArray() is called
with an array that only contains a single value, the edges of the tree
should be drawn in grey.  This bug previously caused such edges to be
drawn in either red or blue.

Change-Id: Ia897109b119029e57e053b368dc910fb87266a76